Restrict click to a version earlier than 8.2.0.
Version 8.2.0 of click introduces incompatible changes to the CliRunner
test class while at the same time increasing the minimum supported
Python version to 3.10. crahsstats-tools still supports Python 3.9, so
we'll use click earlier than 8.2.0 for Python 3.9 and newer versions of
click for newer versions of Python. Since the changes to CliRunner are
backwards-incompatible, we'd need different versions of the test code
for different versions of Python. To simplify things, I'll just pin
click to the old version for now, and we can properly fix this once we
drop support for Python 3.9 later this year.